+++ ballclock/ball_internal_test.go
@@ -5,7 +5,7 @@ import (
 )
 
 func TestBall(t *testing.T) {
-	for i := uint8(0); i < 10; i++ {
+	for i := 0; i < 10; i++ {
 		b := ball{i}
 		if b.id != i {
 			t.Errorf("Test ball failed! %d expected, %d recieved", i, b.id)
@@ -14,7 +14,7 @@ func TestBall(t *testing.T) {
 }
 
 func TestNewBall(t *testing.T) {
-	for i := uint8(0); i < 10; i++ {
+	for i := 0; i < 10; i++ {
 		b := newBall(i)
 		if b.id != i {
 			t.Errorf("Test New() failed! %d expected, %d recieved", i, b.id)
